moved the hasher to string extenstions
