+++ Graffle.FlowSdk.Services/Models/FlowScript.cs
@@ -1,10 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Security.Cryptography;
-using System.Text;
-
 namespace Graffle.FlowSdk.Services.Models
 {
     public class FlowScript
@@ -20,26 +13,5 @@ namespace Graffle.FlowSdk.Services.Models
         //TODO: This needs to remove variables and standardize them
         public string ScriptHash { get; }
 
-
-    }
-
-    //TODO: Fix this and pull it out somewhere proper.
-    public static class ScriptHasher
-    {
-
-        internal static byte[] GetHash(this string inputString)
-        {
-            using (HashAlgorithm algorithm = SHA256.Create())
-                return algorithm.ComputeHash(Encoding.UTF8.GetBytes(inputString));
-        }
-
-        internal static string GetHashString(this string inputString)
-        {
-            StringBuilder sb = new StringBuilder();
-            foreach (byte b in GetHash(inputString))
-                sb.Append(b.ToString("X2"));
-
-            return sb.ToString();
-        }
     }
 }
